# shop-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "InfyOm Laravel Generator APIs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/order-items": {
            "get": {
                "tags": [
                    "OrderItem"
                ],
                "summary": "getOrderItemList",
                "description": "Get all OrderItems",
                "operationId": "0fdb98fdb5ca0daa1bdebfea446aaa18",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/OrderItem"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "OrderItem"
                ],
                "summary": "createOrderItem",
                "description": "Create OrderItem",
                "operationId": "7e765b73ca60025b512b08effcc28375",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderItem"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/OrderItem"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/order-items/{id}": {
            "get": {
                "tags": [
                    "OrderItem"
                ],
                "summary": "getOrderItemItem",
                "description": "Get OrderItem",
                "operationId": "10ef09dd7e6f470f5b4e50f7950496d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of OrderItem",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/OrderItem"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "OrderItem"
                ],
                "summary": "updateOrderItem",
                "description": "Update OrderItem",
                "operationId": "74b59db1d922106dfc2d509c01c6e744",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of OrderItem",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderItem"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/OrderItem"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "OrderItem"
                ],
                "summary": "deleteOrderItem",
                "description": "Delete OrderItem",
                "operationId": "358d3fa86ffd2c83084f378054310f83",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of OrderItem",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "string"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrderItem": {
                "required": [
                    "order_id",
                    "item_id",
                    "quantity",
                    "price_at_purchase"
                ],
                "properties": {
                    "price_at_purchase": {
                        "description": "Price at the time of purchase",
                        "type": "number",
                        "format": "number",
                        "readOnly": false,
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}